LayoutTests/intersection-observer/root-element-deleted-expected.txt

This tests removing the root node of IntersectionObserver. The root node should be eligible for GC.

Test 1: PASS
Test 2: PASS
Test 3: PASS
Test 4: PASS
Test 5: PASS
Test 6: PASS
Test 7: PASS
Test 8: PASS
Test 9: PASS
Test 10: PASS
Test 11: PASS
Test 12: PASS
Test 13: PASS
Test 14: PASS
Test 15: PASS
Test 16: PASS
Test 17: PASS
Test 18: PASS
Test 19: PASS
Test 20: PASS
Node count: PASS
Intersection observer count: PASS
